Added toggleable state to abilities. Made orevision be a toggleable ability - still have to finish it not being dependent on the potion effect. Made energy regen not based on effect buffers.

+++ ponies/abilities/AbilityTeleport.java
@@ -61,7 +61,7 @@ public class AbilityTeleport extends Ability
 	}
 	
 	@Override
-	public int getEnergyCost(EntityPlayer player)
+	public float getEnergyCost(EntityPlayer player)
 	{
 		MovingObjectPosition target = player.rayTrace(maxDistance, 1);
 		if (target == null)
